방패병이 방향전환함에 따라 방패도 따라서 움직이게 만듦 방패의 flip을 위해 NormalEnemy.cs의 Flip함수를 protected로 변경하고 조건문으로 변경 bat의 위치 조정하는 값을 그냥 float로 선언하면 안 되고, double로 선언하고 casting하는 건 되길래 float로 바꿔놓으려던거 원래대로 해놓음

+++ Assets/Scripts/Enemy/Bat.cs
@@ -6,19 +6,19 @@ using System;
 public class Bat : NormalEnemy
 {
 
-    private float coordinate = 0;
+    private double coordinate = 0;
     public ParticleSystem dead;
 
     protected void Roam()
     {
         
         if (coordinate < 2 * Math.PI) {
-            transform.position = new Vector3(transform.position.x + 0.01f, transform.position.y + (Math.Sin(coordinate + 0.05) - Math.Sin(coordinate))/2, transform.position.z);
+            transform.position = new Vector3(transform.position.x + 0.01f, transform.position.y + ((float)Math.Sin(coordinate + 0.05) - (float)Math.Sin(coordinate))/2, transform.position.z);
             coordinate += 0.05;
         }
         else
         {
-            transform.position = new Vector3(transform.position.x - 0.01f, transform.position.y + (Math.Sin(coordinate + 0.05) - Math.Sin(coordinate))/2, transform.position.z);
+            transform.position = new Vector3(transform.position.x - 0.01f, transform.position.y + ((float)Math.Sin(coordinate + 0.05) - (float)Math.Sin(coordinate))/2, transform.position.z);
             coordinate += 0.05;
             if (coordinate > 4 * Math.PI)
             {
